added check if image exist check

+++ frontend/app/api/delete-image.server.ts
@@ -3,6 +3,10 @@ export async function deleteImage(imageId: string, jwt: string) {
   const path = "/api/upload/files/" + imageId;
 
   const url = baseUrl + path;
+  
+  const checkIfImageExists = await fetch(url + imageId);
+
+  if (checkIfImageExists.status === 403) return;
 
   const response = await fetch(url, {
     method: "DELETE",
